LF-4156 move stories to stories folder
